XIN_Picture_Downloader/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Pillow==8.1.2
soupsieve==2.2
